Updated head part with appropriate measurements
I am devastated. I just had this part printed, and realized this morning
that the holes on the top side of the X600 quadcopter frame are not
symmetrical about the line y=x like I had thought. It turns out that my first
print actually had the spacing right for one axis, but my last print had
the y-spacing for one set of holes, and the x-spacing for another set.
It won't fit anywhere.